# legacy-nextjs-app/openapi.yaml
openapi: 3.0.3
info:
  title: KlipCam API
  description: API for KlipCam, a Creator AI platform.
  version: 1.0.0
servers:
  - url: /api
security:
  - ClerkAuth: []
paths:
...
  /stripe/webhook:
    post:
      summary: Stripe webhook
      description: Handles subscription and payment events from Stripe. This endpoint is not called by the frontend client.
      requestBody:
        description: Stripe event object.
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook received and processed.
        '400':
          description: Bad request or invalid signature.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type or code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
      required:
        - error
        - message
...
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token. Include in Authorization header as "Bearer <token>"